Improved I2C Library

+++ I2CDevice.h
@@ -13,14 +13,14 @@ namespace rpiIO{
             ~I2CDevice();
             virtual int open();
             virtual void close();
-            virtual int read(uint8_t *buff, uint16_t len = 1);
-            virtual int write(uint8_t *buff, uint16_t len = 1);
+            virtual int read(uint8_t *buff, int len = 1);
+            virtual int write(uint8_t *buff, int len = 1);
             virtual int read8Register(uint8_t *buff, uint8_t reg);
             virtual int write8Register(uint8_t *buff, uint8_t reg);
             virtual int read16Register(uint16_t *buff, uint8_t reg);
             virtual int write16Register(uint16_t *buff, uint8_t reg);
             /*virtual int readRegisters();*/
-            virtual void dumpRegisters(uint8_t startReg, uint16_t len);
+            virtual void dumpRegisters(uint8_t startReg, int len);
          private:
             int devFile;
             uint8_t devBus;
